requirements.txt
cython
numpy==1.17.0
scipy==1.2.2
Pillow
matplotlib
scikit-image
tensorflow===1.15.0
tensorflow-gpu===1.15.0
keras==2.1.6
opencv-python
h5py==2.9.0
imgaug
IPython[all]
ipython==6.5.0
jupyter-client==5.2.3
jupyter-core==4.4.0
tqdm==4.19.9
git+https://github.com/waleedka/coco.git#subdirectory=PythonAPI
flask
flask-dropzone
flask-ngrok
